// src/types/api/service/models/JournalistAutoCompleteDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * MediaBee Service API
 * MediaBee Service API
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface JournalistAutoCompleteDto
 */
export interface JournalistAutoCompleteDto {
    /**
     * 
     * @type {number}
     * @memberof JournalistAutoCompleteDto
     */
    journalistId?: number;
    /**
     * 
     * @type {string}
     * @memberof JournalistAutoCompleteDto
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof JournalistAutoCompleteDto
     */
    department?: string;
    /**
     * 
     * @type {string}
     * @memberof JournalistAutoCompleteDto
     */
    position?: string;
    /**
     * 
     * @type {number}
     * @memberof JournalistAutoCompleteDto
     */
    mediaId?: number;
    /**
     * 
     * @type {string}
     * @memberof JournalistAutoCompleteDto
     */
    mediaName?: string;
    /**
     * 
     * @type {string}
     * @memberof JournalistAutoCompleteDto
     */
    email?: string;
}

/**
 * Check if a given object implements the JournalistAutoCompleteDto interface.
 */
export function instanceOfJournalistAutoCompleteDto(value: object): value is JournalistAutoCompleteDto {
    return true;
}

export function JournalistAutoCompleteDtoFromJSON(json: any): JournalistAutoCompleteDto {
    return JournalistAutoCompleteDtoFromJSONTyped(json, false);
}

export function JournalistAutoCompleteDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): JournalistAutoCompleteDto {
    if (json == null) {
        return json;
    }
    return {
        
        'journalistId': json['journalistId'] == null ? undefined : json['journalistId'],
        'name': json['name'] == null ? undefined : json['name'],
        'department': json['department'] == null ? undefined : json['department'],
        'position': json['position'] == null ? undefined : json['position'],
        'mediaId': json['mediaId'] == null ? undefined : json['mediaId'],
        'mediaName': json['mediaName'] == null ? undefined : json['mediaName'],
        'email': json['email'] == null ? undefined : json['email'],
    };
}

export function JournalistAutoCompleteDtoToJSON(value?: JournalistAutoCompleteDto | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'journalistId': value['journalistId'],
        'name': value['name'],
        'department': value['department'],
        'position': value['position'],
        'mediaId': value['mediaId'],
        'mediaName': value['mediaName'],
        'email': value['email'],
    };
}